Bg video added
Splash screen added with autoplaying video background. Next steps: header/nav overlay + react-scroll.
