+++ src/App.js
@@ -1,10 +1,9 @@
 import { Routes, Route } from 'react-router-dom';
-import { Header, Landing, About, Footer, Error } from './components/exports.js';
+import { Landing, About, Footer, Error } from './components/exports.js';
 
 function App() {
     return (
         <>
-            <Header />
             <Routes>
                 <Route path="/" element={<Landing />} />
                 <Route path="/about" element={<About />} />
