+++ src/components/exports.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import Footer from './footer-banner/index.js';
-import Header from './header-banner/index.js';
 
 //pages
 import Landing from './pages/landing/index.js';
 import About from './pages/about/index.js';
 import Error from './pages/error/index.js';
 
-export { Footer, Header, Landing, About, Error };
+export { Footer, Landing, About, Error };
