+++ src/components/pages/about/index.js
@@ -1,12 +1,11 @@
 import React from 'react';
-
 const About = () => {
     return (
         <main>
-            <p>about</p>
-            <p>under development</p>
+            <p>About Us</p>
+            <p>The link you chose is not working. We're working on it!</p>
         </main>
     );
 };
 
-export default About;
+export default Error;
