+++ src/components/pages/landing/index.js
@@ -1,10 +1,14 @@
 import React from 'react';
+import Header from './header-banner/index.js';
 import ProjInfo from './proj-info';
 import Missions from './missions';
 import News from './news';
 import Papers from './papers';
 import Media from './media';
 import Teams from './teams';
+
+//assets
+import videoBg from '../../../assets/banner_video.mp4';
 
 const Landing = () => {
     const maxWidth = {
@@ -15,8 +19,15 @@ const Landing = () => {
     };
     return (
         <>
-            <main className="antialiased">
-                <h1>Landing</h1>
+            <main className="antialiased w-full min-h-screen">
+                <div id="video">
+                    <div id="overlay" className="absolute top-0 left-0 w-full h-full bg-black bg-opacity-50"></div>
+                    <video src={videoBg} autoPlay loop muted className="w-full h-screen object-cover" />
+                    <div id="videoContent" className="absolute w-full h-full top-0 flex flex-col justify-center items-center text-neutral-100">
+                        <h1>Welcome</h1>
+                        <p>Scroll Down</p>
+                    </div>
+                </div>
                 <ProjInfo />
                 <Missions />
                 <News />
@@ -24,301 +35,6 @@ const Landing = () => {
                 <Media />
                 <Teams />
             </main>
-            <section id="test">
-                {/* <!-- Navbar (sit on top) --> */}
-                <div className="w3-top">
-                    <div className="w3-bar" id="myNavbar">
-                        <a
-                            className="w3-bar-item w3-button w3-hover-black w3-hide-medium w3-hide-large w3-right"
-                            href="javascript:void(0);"
-                            onclick="toggleFunction()"
-                            title="Toggle Navigation Menu"
-                        >
-                            <i className="fa fa-bars"></i>
-                        </a>
-                        <a href="#home" className="w3-bar-item w3-button">
-                            HOME
-                        </a>
-                        <a href="#about" className="w3-bar-item w3-button w3-hide-small">
-                            <i className="fa fa-user"></i> ABOUT
-                        </a>
-                        <a href="#portfolio" className="w3-bar-item w3-button w3-hide-small">
-                            <i className="fa fa-th"></i> PORTFOLIO
-                        </a>
-                        <a href="#contact" className="w3-bar-item w3-button w3-hide-small">
-                            <i className="fa fa-envelope"></i> CONTACT
-                        </a>
-                        <a href="#" className="w3-bar-item w3-button w3-hide-small w3-right w3-hover-red">
-                            <i className="fa fa-search"></i>
-                        </a>
-                    </div>
-
-                    {/* <!-- Navbar on small screens --> */}
-                    <div id="navDemo" className="w3-bar-block w3-white w3-hide w3-hide-large w3-hide-medium">
-                        <a href="#about" className="w3-bar-item w3-button" onclick="toggleFunction()">
-                            ABOUT
-                        </a>
-                        <a href="#portfolio" className="w3-bar-item w3-button" onclick="toggleFunction()">
-                            PORTFOLIO
-                        </a>
-                        <a href="#contact" className="w3-bar-item w3-button" onclick="toggleFunction()">
-                            CONTACT
-                        </a>
-                        <a href="#" className="w3-bar-item w3-button">
-                            SEARCH
-                        </a>
-                    </div>
-                </div>
-
-                {/* <!-- First Parallax Image with Logo Text --> */}
-                <div className="bgimg-1 w3-display-container w3-opacity-min" id="home">
-                    <div className="w3-display-middle" style={{ whiteSpace: 'nowrap' }}>
-                        <span className="w3-center w3-padding-large w3-black w3-xlarge w3-wide w3-animate-opacity">
-                            MY <span className="w3-hide-small">WEBSITE</span> LOGO
-                        </span>
-                    </div>
-                </div>
-
-                {/* <!-- Container (About Section) --> */}
-                <div className="w3-content w3-container w3-padding-64" id="about">
-                    <h3 className="w3-center">ABOUT ME</h3>
-                    <p className="w3-center">
-                        <em>I love photography</em>
-                    </p>
-                    <p>
-                        We have created a fictional "personal" website/blog, and our fictional character is a hobby photographer. Lorem ipsum dolor sit amet,
-                        consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-                        exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse
-                        cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim
-                        id est laborum consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
-                        quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-                    </p>
-                    <div className="w3-row">
-                        <div className="w3-col m6 w3-center w3-padding-large">
-                            <p>
-                                <b>
-                                    <i className="fa fa-user w3-margin-right"></i>My Name
-                                </b>
-                            </p>
-                            <br></br>
-                            <img
-                                src="/w3images/avatar_hat.jpg"
-                                className="w3-round w3-image w3-opacity w3-hover-opacity-off"
-                                alt="Photo of Me"
-                                width="500"
-                                height="333"
-                            />
-                        </div>
-
-                        {/* <!-- Hide this text on small devices --> */}
-                        <div className="w3-col m6 w3-hide-small w3-padding-large">
-                            <p>
-                                Welcome to my website. I am lorem ipsum consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna
-                                aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute
-                                irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat
-                                non proident, sunt in culpa qui officia deserunt mollit anim id est laborum consectetur adipiscing elit, sed do eiusmod tempor
-                                incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip
-                                ex ea commodo consequat.
-                            </p>
-                        </div>
-                    </div>
-                    <p className="w3-large w3-center w3-padding-16">Im really good at:</p>
-                    <p className="w3-wide">
-                        <i className="fa fa-camera"></i>Photography
-                    </p>
-                    <div className="w3-light-grey">
-                        <div className="w3-container w3-padding-small w3-dark-grey w3-center" style={{ width: '90%' }}>
-                            90%
-                        </div>
-                    </div>
-                    <p className="w3-wide">
-                        <i className="fa fa-laptop"></i>Web Design
-                    </p>
-                    <div className="w3-light-grey">
-                        <div className="w3-container w3-padding-small w3-dark-grey w3-center" style={{ width: '85%' }}>
-                            85%
-                        </div>
-                    </div>
-                    <p className="w3-wide">
-                        <i className="fa fa-photo"></i>Photoshop
-                    </p>
-                    <div className="w3-light-grey">
-                        <div className="w3-container w3-padding-small w3-dark-grey w3-center" style={{ width: '75%' }}>
-                            75%
-                        </div>
-                    </div>
-                </div>
-
-                <div className="w3-row w3-center w3-dark-grey w3-padding-16">
-                    <div className="w3-quarter w3-section">
-                        <span className="w3-xlarge">14+</span>
-                        <br></br>
-                        Partners
-                    </div>
-                    <div className="w3-quarter w3-section">
-                        <span className="w3-xlarge">55+</span>
-                        <br></br>
-                        Projects Done
-                    </div>
-                    <div className="w3-quarter w3-section">
-                        <span className="w3-xlarge">89+</span>
-                        <br></br>
-                        Happy Clients
-                    </div>
-                    <div className="w3-quarter w3-section">
-                        <span className="w3-xlarge">150+</span>
-                        <br></br>
-                        Meetings
-                    </div>
-                </div>
-
-                {/* <!-- Second Parallax Image with Portfolio Text --> */}
-                <div className="bgimg-2 w3-display-container w3-opacity-min">
-                    <div className="w3-display-middle">
-                        <span className="w3-xxlarge w3-text-white w3-wide">PORTFOLIO</span>
-                    </div>
-                </div>
-
-                {/* <!-- Container (Portfolio Section) --> */}
-                <div className="w3-content w3-container w3-padding-64" id="portfolio">
-                    <h3 className="w3-center">MY WORK</h3>
-                    <p className="w3-center">
-                        <em>
-                            Here are some of my latest lorem work ipsum tipsum.<br></br> Click on the images to make them bigger
-                        </em>
-                    </p>
-                    <br></br>
-
-                    {/* <!-- Responsive Grid. Four columns on tablets, laptops and desktops. Will stack on mobile devices/small screens (100% width) --> */}
-                    <div className="w3-row-padding w3-center">
-                        <div className="w3-col m3">
-                            <img
-                                src="../../../w3images/p1.jpg"
-                                style={maxWidth}
-                                onclick="onClick(this)"
-                                className="w3-hover-opacity"
-                                alt="The mist over the mountains"
-                            />
-                        </div>
-
-                        <div className="w3-col m3">
-                            <img src="../../../w3images/p2.jpg" style={maxWidth} onclick="onClick(this)" className="w3-hover-opacity" alt="Coffee beans" />
-                        </div>
-
-                        <div className="w3-col m3">
-                            <img src="../../../w3images/p3.jpg" style={maxWidth} onclick="onClick(this)" className="w3-hover-opacity" alt="Bear closeup" />
-                        </div>
-
-                        <div className="w3-col m3">
-                            <img src="../../../w3images/p4.jpg" style={maxWidth} onclick="onClick(this)" className="w3-hover-opacity" alt="Quiet ocean" />
-                        </div>
-                    </div>
-
-                    <div className="w3-row-padding w3-center w3-section">
-                        <div className="w3-col m3">
-                            <img src="../../../w3images/p5.jpg" style={maxWidth} onclick="onClick(this)" className="w3-hover-opacity" alt="The mist" />
-                        </div>
-
-                        <div className="w3-col m3">
-                            <img
-                                src="../../../w3images/p6.jpg"
-                                style={maxWidth}
-                                onclick="onClick(this)"
-                                className="w3-hover-opacity"
-                                alt="My beloved typewriter"
-                            />
-                        </div>
-
-                        <div className="w3-col m3">
-                            <img src="../../../w3images/p7.jpg" style={maxWidth} onclick="onClick(this)" className="w3-hover-opacity" alt="Empty ghost train" />
-                        </div>
-
-                        <div className="w3-col m3">
-                            <img src="../../../w3images/p8.jpg" style={maxWidth} onclick="onClick(this)" className="w3-hover-opacity" alt="Sailing" />
-                        </div>
-                        <button className="w3-button w3-padding-large w3-light-grey" style={marginTop64}>
-                            LOAD MORE
-                        </button>
-                    </div>
-                </div>
-
-                {/* <!-- Modal for full size images on click--> */}
-                <div id="modal01" className="w3-modal w3-black" onclick="this.style.display='none'">
-                    <span className="w3-button w3-large w3-black w3-display-topright" title="Close Modal Image">
-                        <i className="fa fa-remove"></i>
-                    </span>
-                    <div className="w3-modal-content w3-animate-zoom w3-center w3-transparent w3-padding-64">
-                        <img id="img01" className="w3-image" />
-                        <p id="caption" className="w3-opacity w3-large"></p>
-                    </div>
-                </div>
-
-                {/* <!-- Third Parallax Image with Portfolio Text --> */}
-                <div className="bgimg-3 w3-display-container w3-opacity-min">
-                    <div className="w3-display-middle">
-                        <span className="w3-xxlarge w3-text-white w3-wide">CONTACT</span>
-                    </div>
-                </div>
-
-                {/* <!-- Container (Contact Section) --> */}
-                <div className="w3-content w3-container w3-padding-64" id="contact">
-                    <h3 className="w3-center">WHERE I WORK</h3>
-                    <p className="w3-center">
-                        <em>I'd love your feedback!</em>
-                    </p>
-
-                    <div className="w3-row w3-padding-32 w3-section">
-                        <div className="w3-col m4 w3-container">
-                            <img src="./../../../w3images/map.jpg" className="w3-image w3-round" style={maxWidth} />
-                        </div>
-                        <div className="w3-col m8 w3-panel">
-                            <div className="w3-large w3-margin-bottom">
-                                <i className="fa fa-map-marker fa-fw w3-hover-text-black w3-xlarge w3-margin-right"></i> Chicago, US<br></br>
-                                <i className="fa fa-phone fa-fw w3-hover-text-black w3-xlarge w3-margin-right"></i> Phone: +00 151515<br></br>
-                                <i className="fa fa-envelope fa-fw w3-hover-text-black w3-xlarge w3-margin-right"></i> Email: mail@mail.com<br></br>
-                            </div>
-                            <p>
-                                Swing by for a cup of <i className="fa fa-coffee"></i>, or leave me a note:
-                            </p>
-                            <form action="/action_page.php" target="_blank">
-                                <div className="w3-row-padding" style={{ margin: '0 -16px 8px -16px' }}>
-                                    <div className="w3-half">
-                                        <input className="w3-input w3-border" type="text" placeholder="Name" required name="Name" />
-                                    </div>
-                                    <div className="w3-half">
-                                        <input className="w3-input w3-border" type="text" placeholder="Email" required name="Email" />
-                                    </div>
-                                </div>
-                                <input className="w3-input w3-border" type="text" placeholder="Message" required name="Message" />
-                                <button className="w3-button w3-black w3-right w3-section" type="submit">
-                                    <i className="fa fa-paper-plane"></i> SEND MESSAGE
-                                </button>
-                            </form>
-                        </div>
-                    </div>
-                </div>
-
-                {/* <!-- Footer --> */}
-                <footer className="w3-center w3-black w3-padding-64 w3-opacity w3-hover-opacity-off">
-                    <a href="#home" className="w3-button w3-light-grey">
-                        <i className="fa fa-arrow-up w3-margin-right"></i>To the top
-                    </a>
-                    <div className="w3-xlarge w3-section">
-                        <i className="fa fa-facebook-official w3-hover-opacity"></i>
-                        <i className="fa fa-instagram w3-hover-opacity"></i>
-                        <i className="fa fa-snapchat w3-hover-opacity"></i>
-                        <i className="fa fa-pinterest-p w3-hover-opacity"></i>
-                        <i className="fa fa-twitter w3-hover-opacity"></i>
-                        <i className="fa fa-linkedin w3-hover-opacity"></i>
-                    </div>
-                    <p>
-                        Powered by{' '}
-                        <a href="https://www.w3schools.com/w3css/default.asp" title="W3.CSS" target="_blank" className="w3-hover-text-green">
-                            w3.css
-                        </a>
-                    </p>
-                </footer>
-            </section>
         </>
     );
 };
