+++ src/index.css
@@ -2,48 +2,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-/* Create a Parallax Effect */
-.bgimg-1,
-.bgimg-2,
-.bgimg-3 {
-    background-attachment: fixed;
-    background-position: center;
-    background-repeat: no-repeat;
-    background-size: cover;
-}
-
-/* First image (Logo. Full height) */
-.bgimg-1 {
-    background-image: url('./w3images/parallax1.jpg');
-    min-height: 100%;
-}
-
-/* Second image (Portfolio) */
-.bgimg-2 {
-    background-image: url('./w3images/parallax2.jpg');
-    min-height: 400px;
-}
-
-/* Third image (Contact) */
-.bgimg-3 {
-    background-image: url('./w3images/parallax3.jpg');
-    min-height: 400px;
-}
-
-.w3-wide {
-    letter-spacing: 10px;
-}
-.w3-hover-opacity {
-    cursor: pointer;
-}
-
-/* Turn off parallax scrolling for tablets and phones */
-@media only screen and (max-device-width: 1600px) {
-    .bgimg-1,
-    .bgimg-2,
-    .bgimg-3 {
-        background-attachment: scroll;
-        min-height: 400px;
-    }
-}
